kafka: move kafka probes into server library
