+++ src/v/kafka/server/connection_context.cc
@@ -19,7 +19,6 @@
 #include "cluster/types.h"
 #include "config/configuration.h"
 #include "kafka/protocol/sasl_authenticate.h"
-#include "kafka/sasl_probe.h"
 #include "kafka/server/handlers/fetch.h"
 #include "kafka/server/handlers/handler_interface.h"
 #include "kafka/server/handlers/produce.h"
@@ -28,6 +27,7 @@
 #include "kafka/server/quota_manager.h"
 #include "kafka/server/request_context.h"
 #include "kafka/server/response.h"
+#include "kafka/server/sasl_probe.h"
 #include "kafka/server/server.h"
 #include "kafka/server/snc_quota_manager.h"
 #include "net/exceptions.h"
