+++ src/v/kafka/server/group.h
@@ -18,10 +18,10 @@
 #include "container/chunked_hash_map.h"
 #include "container/fragmented_vector.h"
 #include "features/feature_table.h"
-#include "kafka/group_probe.h"
 #include "kafka/protocol/fwd.h"
 #include "kafka/protocol/offset_commit.h"
 #include "kafka/server/group_metadata.h"
+#include "kafka/server/group_probe.h"
 #include "kafka/server/member.h"
 #include "model/fundamental.h"
 #include "model/record.h"
