+++ src/v/kafka/server/group_probe.h
