+++ src/v/kafka/server/server.h
@@ -14,17 +14,17 @@
 #include "cluster/fwd.h"
 #include "config/configuration.h"
 #include "features/feature_table.h"
-#include "kafka/latency_probe.h"
 #include "kafka/protocol/types.h"
-#include "kafka/read_distribution_probe.h"
-#include "kafka/sasl_probe.h"
 #include "kafka/server/connection_context.h"
 #include "kafka/server/fetch_metadata_cache.hh"
 #include "kafka/server/fetch_session_cache.h"
 #include "kafka/server/fwd.h"
 #include "kafka/server/handlers/fetch/replica_selector.h"
 #include "kafka/server/handlers/handler_probe.h"
+#include "kafka/server/latency_probe.h"
 #include "kafka/server/queue_depth_monitor.h"
+#include "kafka/server/read_distribution_probe.h"
+#include "kafka/server/sasl_probe.h"
 #include "metrics/metrics.h"
 #include "net/server.h"
 #include "pandaproxy/schema_registry/fwd.h"
